Move HashAlgorithmOperations to the MS.ANC.Razor.Utilties namespace
