+++ src/Compiler/Microsoft.AspNetCore.Razor.Language/src/StringSourceDocument.cs
@@ -5,6 +5,7 @@
 
 using System;
 using System.Text;
+using Microsoft.AspNetCore.Razor.Utilities;
 
 namespace Microsoft.AspNetCore.Razor.Language;
